# restAPI/src/main/swagger/openapi.yaml
openapi: 3.0.0
info:
  description: This is the API for the UI's access to data storage
  version: 1.0.0
  title: AIDA Store REST API
servers:
  - url: "{protocol}://{host}:{port}/{basePath}"
    variables:
      protocol:
        enum:
          - http
          - https
        default: http
      host:
        default: localhost
      port:
        default: "8008"
      basePath:
        default: api
...
paths:
...
  /relation:
    get:
      tags:
        - Relation
      summary: Return a single relation
      description: Return relation by id
      operationId: relation
      parameters:
        - name: id
          in: query
          description: ID of relation to retrieve
          required: true
          schema:
            type: string
          example: http://www.lti.cs.cmu.edu/aida/opera/corpora/eval/cluster-relation-instance-JC002YEM4-r202010162240-43
        - $ref: "#/components/parameters/optionalGraphParam"
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "400":
          description: Invalid parameters
        "404":
          description: API not found
...
  /roles:
    get:
      tags:
        - Roles
      summary: Return roles
      description: Return all argument roles and fillers for a particular node
      operationId: roles
      parameters:
        - name: id
          in: query
          description: ID of relation to retrieve
          required: true
          schema:
            type: string
          example: http://www.lti.cs.cmu.edu/aida/opera/corpora/eval/event-instance-IC001V5AJ-r202010162251-28
        - $ref: "#/components/parameters/optionalGraphParam"
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Role"
        "400":
          description: Invalid parameters
        "404":
          description: API not found
...
  /entity/element:
    get:
      tags:
        - Entity
      summary: Return entity element
      description: Return an entity element by id
      operationId: entityElement
      parameters:
        - name: id
          in: query
          description: ID of entity element to retrieve
          required: true
          schema:
            type: string
          example: http://www.lti.cs.cmu.edu/aida/opera/corpora/eval/cross_doc_coref-opera-entity-combo-r202101220414-7659-entity
        - $ref: "#/components/parameters/optionalGraphParam"
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EntityElement"
        "400":
          description: Invalid parameters
        "404":
          description: API not found
...
components:
  schemas:
...
    Event:
      type: object
      properties:
        cluster:
          type: string
        handle:
          type: string
        category:
          type: string
        prototype:
          $ref: "#/components/schemas/Member"
        members:
          type: array
          items:
            $ref: "#/components/schemas/Member"
    Member:
      allOf:
        - $ref: "#/components/schemas/JustifiedNode"
        - type: object
          properties:
            arguments:
              type: array
              items:
                $ref: "#/components/schemas/Role"
...
    Role:
      type: object
      properties:
        role:
          type: string
        fillers:
          type: array
          items:
            $ref: "#/components/schemas/Entity"
    JustifiedNode:
      type: object
      discriminator:
        propertyName: objectType
      properties:
        id:
          type: string
        handle:
          type: string
        objectType:
          type: string
          enum: ["Entity", "Member", "Element", "EntityMember", "EntityElement", "JustifiedNode"]
        types:
          type: array
          items:
            type: string
        docs:
          type: array
          items:
            $ref: "#/components/schemas/DocObject"
    Entity:
      allOf:
        - $ref: "#/components/schemas/JustifiedNode"
        - type: object
          properties:
            names:
              type: array
              items:
                type: string
            category:
              type: string
            clusterIds:
              type: array
              items:
                type: string
...
    EntityMember:
      allOf:
        - $ref: "#/components/schemas/Entity"
        - type: object
          properties:
            connections:
              type: array
              items:
                $ref: "#/components/schemas/ReverseRole"
    EntityElement:
      allOf:
        - $ref: "#/components/schemas/EntityMember"
        - type: object
          properties:
            clusters:
              type: array
              items:
                $ref: "#/components/schemas/Entity"
    ReverseRole:
      type: object
      properties:
        role:
          type: string
        category:
          type: string
        connections:
          type: array
          items:
            $ref: "#/components/schemas/JustifiedNode"
...
    DocObject:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        contentDate:
          type: string
        downloadDate:
          type: string
...
  parameters:
...
    optionalGraphParam:
      name: graph
      in: query
      description: Which graph to query in
      schema:
        type: string
      example: https://www.nextcentury.com/TA2/OPERA-20210215/OPERA-20210117
...